+++ juriscraper/opinions/united_states/attorney_general/__init__.py
@@ -1,0 +1,8 @@
+__all__ = [
+    "calag",
+    "mdag",
+    "minnag",
+    "nyag",
+    "oklaag",
+    "texag",
+]